% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_mni_atlas.R
\name{download_mni_atlas}
\alias{download_mni_atlas}
\title{Download MNI Atlas Zip file}
\usage{
download_mni_atlas(atlas = names(nitrc::mni_atlas_links), out_dir = ".",
  verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{atlas}{Name of atlas to download.  See \code{\link{mni_atlas_links}}}

\item{out_dir}{Output directory, directory to download izp file}

\item{verbose}{Print diagnostic messages}

\item{overwrite}{Should the file be overwritten if it exists}
}
\value{
List of 2 elements, return of \code{GET}, and output filename
}
\description{
Downloads a single MNI atlas from the MNI website
(http://nist.mni.mcgill.ca/?page_id=714)
}
\examples{
download_mni_atlas("mni305", out_dir = tempdir())
}

