% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nitrc_ir.R
\name{download_nitrc_ir}
\alias{download_nitrc_ir}
\title{Download NITRC Image Repository}
\usage{
download_nitrc_ir(study = c("ixi", "pd", "fcon_1000", "candi"), username,
  password, out_dir = ".", verbose = TRUE)
}
\arguments{
\item{study}{study to be parsed, code see \code{\link{nitrc_ir_study}}}

\item{username}{Username for authentication of NITRC}

\item{password}{Password for authentication of NITRC}

\item{out_dir}{Output directory}

\item{verbose}{Print diagnostic messages}
}
\value{
Vector of names of the output files
}
\description{
Downloads an entire NITRC Image Repository,
wrapper function for \code{\link{download_nitrc_ir_url}}
}

