% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_nitrc_ir_url.R
\name{download_nitrc_ir_url}
\alias{download_nitrc_ir_url}
\title{Download NITRC Single Image Repository URL}
\usage{
download_nitrc_ir_url(url, username, password, out_dir = ".",
  verbose = TRUE, ...)
}
\arguments{
\item{url}{URL of file to download}

\item{username}{Username for authentication of NITRC}

\item{password}{Password for authentication of NITRC}

\item{out_dir}{Output directory}

\item{verbose}{Print diagnostic messages}

\item{...}{additional options to pass to \code{\link{download_nitrc_url}}}
}
\value{
Vector of names of the output files
}
\description{
Downloads a single URL from a NITRC Image Repository,
wrapper function for \code{\link{download_nitrc_url}}
}

