% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mni_atlas_links.R
\docType{data}
\name{mni_atlas_links}
\alias{mni_atlas_links}
\title{MNI atlas Links}
\format{An object of class \code{character} of length 29.}
\usage{
mni_atlas_links
}
\description{
A named vector of links to the MNI atlases from
http://nist.mni.mcgill.ca/?page_id=714
}
\keyword{datasets}
