% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrc_documents.R
\name{nitrc_documents}
\alias{nitrc_documents}
\title{NITRC Documents}
\usage{
nitrc_documents(..., long = FALSE)
}
\arguments{
\item{...}{arguments passed to \code{\link{nitrc_remote_data}}}

\item{long}{passed to \code{\link{parse_nitrc_docs}}}
}
\value{
\code{data.frame}, same as \code{\link{parse_nitrc_docs}}
}
\description{
Wraps \code{\link{nitrc_remote_data}} and
\code{\link{parse_nitrc_docs}} to return NITRC document information from
the projects
}
\seealso{
\code{\link{parse_nitrc_docs}}
}
