% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrc_ir_study.R
\name{nitrc_ir_study}
\alias{nitrc_ir_study}
\title{NITRC Image Repository Study}
\usage{
nitrc_ir_study(study = c("ixi", "pd", "fcon_1000", "candi"))
}
\arguments{
\item{study}{study to be parsed
ixi = IXI,
pd = High-quality diffusion-weighted imaging of Parkinson's disease
fcon_1000 = 1000 Functional Connectomes
candi = CANDI Share: Schizophrenia Bulletin 2008}
}
\value{
List of a data.frame of the files to download and the demographics
}
\description{
Parses the necessary information for the specified NITRC
Image Repository (IR) Study
}
\examples{
x = nitrc_ir_study("ixi")
}

