% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrc_packages.R
\name{nitrc_packages}
\alias{nitrc_packages}
\title{NITRC Packages}
\usage{
nitrc_packages(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{nitrc_remote_data}}}
}
\value{
\code{data.frame}, same as \code{\link{parse_nitrc_packs}}
}
\description{
Wraps \code{\link{nitrc_remote_data}} and
\code{\link{parse_nitrc_packs}} to return NITRC package information from
the projects
}
\seealso{
\code{\link{parse_nitrc_packs}}
}
