% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrc_remote_data.R
\name{nitrc_remote_data}
\alias{nitrc_remote_data}
\title{Download NITRC Remote Data Sources}
\usage{
nitrc_remote_data(spec = c("all", "", "local", "incf"))
}
\arguments{
\item{spec}{specification of data to extract.  Default is All.  See
\url{http://www.nitrc.org/plugins/mwiki/index.php/nitrc:NITRC_Remote_Data_Access}
for more information}
}
\value{
A list from \code{\link[jsonlite]{fromJSON}} of the parsed content
returned
}
\description{
Downloads the JSON object from the NITRC Remote Data Sources,
and parses it to a list.
}
\examples{
res = nitrc_remote_data(spec = "589")
}
