% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nitrc_docs.R
\name{parse_nitrc_docs}
\alias{parse_nitrc_docs}
\title{Parse Documents from NITRC List}
\usage{
parse_nitrc_docs(proj, long = FALSE)
}
\arguments{
\item{proj}{List of projects returned from \code{\link{nitrc_remote_data}}}

\item{long}{logical indicator if the \code{data.frame} should be returned
long or wide}
}
\value{
\code{data.frame}
}
\description{
Parses the list of projects from \code{\link{nitrc_remote_data}} to a
\code{data.frame} of documents
}
\examples{
\dontrun{
js = nitrc_remote_data(spec = "all")
proj = js$projects
doc_df = parse_nitrc_docs(proj)
}
}
