% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nitrc_packs.R
\name{parse_nitrc_packs}
\alias{parse_nitrc_packs}
\title{Parse Packages from NITRC List}
\usage{
parse_nitrc_packs(proj)
}
\arguments{
\item{proj}{List of projects returned from \code{\link{nitrc_remote_data}}}
}
\value{
\code{data.frame}
}
\description{
Parses the list of projects from \code{\link{nitrc_remote_data}} to a
\code{data.frame} of packages
}
\examples{
\dontrun{
js = nitrc_remote_data(spec = "all")
proj = js$projects
doc_df = parse_nitrc_packs(proj)
projects = parse_nitrc_proj(proj, add_attr = TRUE)
dproj = projects[ tolower(projects$attr_name) \%in\% "data", ]
data_doc = doc_df[ doc_df$proj_id \%in\% dproj$proj_id, ]
data_doc$type = trimws(data_doc$type)
data_doc$type = trimws(gsub("^Source", "", data_doc$type))
}
}
