% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nitrc_proj.R
\name{parse_nitrc_proj}
\alias{parse_nitrc_proj}
\title{Parse Project Information from NITRC List}
\usage{
parse_nitrc_proj(proj, add_attr = FALSE)
}
\arguments{
\item{proj}{List of projects returned from \code{\link{nitrc_remote_data}}}

\item{add_attr}{Should attributes be added (this will replicate multiple
records to the same ID)}
}
\value{
\code{data.frame}
}
\description{
Parses the list of projects from \code{\link{nitrc_remote_data}} to a
\code{data.frame} of projects with attributes
}
\examples{
\dontrun{
js = nitrc_remote_data(spec = "all")
proj = js$projects
doc_df = parse_nitrc_proj(proj)
}
}
