% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcat_reader.R
\name{xcat_reader}
\alias{xcat_reader}
\title{XML Category Reader from NITRC}
\usage{
xcat_reader(xml_file, all_mods)
}
\arguments{
\item{xml_file}{XML file from NITRC}

\item{all_mods}{types of modalities in the data}
}
\value{
\code{data.frame} of URL, ID, and mods
}
\description{
Reads XML Category file from NITRC
}
